requirements.txt
flask
gdown
opencv-python-headless
tensorflow-cpu
numpy
gunicorn
